% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatmask-package.R
\docType{package}
\name{spatmask-package}
\alias{spatmask}
\alias{spatmask-package}
\title{spatmask: Spatial Geomasking and Anonymization Tools}
\description{
Tools for spatial geomasking and anonymization of geographic data.
spatmask provides methods to protect privacy in spatially referenced datasets
while preserving analytical utility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JDConejeros/spatmask}
  \item Report bugs at \url{https://github.com/JDConejeros/spatmask/issues}
}

}
\author{
\strong{Maintainer}: José Daniel Conejeros \email{jo.conejeros@gmail.com} (\href{https://orcid.org/0000-0003-3402-4361}{ORCID})

}
\keyword{internal}
